Restructure api middlewares
