+++ src/api/polls/poll.routes.ts
@@ -1,7 +1,8 @@
 import { Router } from 'express';
 import PollController from './poll.controller';
 import { validate } from 'express-validation';
-import { authRejoin, validateCreatePoll, validateJoinPoll } from './poll.middleware';
+import { authRejoin } from './middlewares/auth';
+import { validateCreatePoll, validateJoinPoll } from './middlewares/validate';
 
 const router: Router = Router();
